+++ packages/auto_interop_generator/test/unit/generators/dart_generator_test.dart
@@ -244,7 +244,8 @@ void main() {
               ],
               returnType: UtsType.voidType(),
             ),
-          ], types: [
+          ],
+          types: [
             UtsClass(
               name: 'Options',
               kind: UtsClassKind.dataClass,